refactor(#138): move lint_feature_coverage into linting.py (module-split)
Continue the module-split: the feature-coverage completeness check
(lint_feature_coverage, 100 lines) moves out of make_drawing.py into linting.py,
joining CoverageState (the coverage signal it reads). Drawing.lint() keeps
calling it via a re-import; it stays public (__init__ re-exports it, tests import
it from make_drawing — both keep working).

Pure code move — no behaviour change:
- linting.py depends only on _core (_DIAM_RE, _fmt) and build123d_drafting
  (analyse_cylinders/feature_diameters/find_holes, LintIssue, TitleBlock); no
  make_drawing import, so no cycle.
- _suggest_fix is deliberately NOT moved yet: it shares _QUOTED_RE with the repair
  code, so it rides with the repair.py extraction (relocating _QUOTED_RE to _core)
  rather than forcing a cross-module constant now. Keeps this diff clean.
- The golden gate passes against the committed snapshots with no regeneration, and
  the lint_feature_coverage tests pass.
