+++ CLAUDE.md
@@ -42,10 +42,11 @@ The dependency graph is a DAG: the leaf modules `layout.py`, `registry.py`,
   identity/ownership/pins/build-issues (#138 / ADR 0005, Step 2). `Drawing`
   delegates here and keeps the render list; `_named`/`_anno_view`/`_pinned`/
   `_build_issues` remain `Drawing` properties during the migration.
-- **`linting.py`** — `CoverageState`: the single owner of the lint-side coverage
-  signal (pattern callouts, patterned holes, dropped callout diameters; #138 /
-  ADR 0005, Step 3). The designated future home for `lint_feature_coverage` /
-  `_suggest_fix`; for now it owns just the state they read.
+- **`linting.py`** — the lint module (#138 / ADR 0005): `lint_feature_coverage`
+  (the feature-coverage completeness check) and `CoverageState` (the coverage
+  signal it reads — pattern callouts, patterned holes, dropped diameters).
+  Depends only on `_core` + build123d_drafting. (`_suggest_fix` follows with the
+  repair extraction — it shares `_QUOTED_RE` with the repair code.)
 - **`fonts.py`** — vendored, path-pinned IBM Plex fonts for deterministic
   cross-platform layout (ADR 0006).
 - **`export.py`** — SVG/DXF/PDF export + post-processing (page-size fix,
